refactor: パーツ編集の構成コンポーネントを別ファイルに独立
